add throw statements that throw exceptions instead of assert()

+++ include/Hatrix/classes/Matrix.h
@@ -129,8 +129,8 @@ class Matrix {
 
   Matrix block_ranks(int64_t nblocks, double accuracy) const;
 
-  Matrix swap_rows(const std::vector<int>& row_indices);
-  Matrix swap_cols(const std::vector<int>& col_indices);
+  Matrix swap_rows(const std::vector<int64_t>& row_indices);
+  Matrix swap_cols(const std::vector<int64_t>& col_indices);
 };
 
 }  // namespace Hatrix
